tests/DCPS/RtpsRelay/Unit does not conform to guidelines
Problem
-------

Unit tests under tests/DCPS/RtpsRelay/Unit do not conform to
docs/internal/unit_tests.rst.

Solution
--------

Move unit tests according to guidlines.
