--- tests/DCPS/RtpsRelay/Unit/.gitignore
@@ -1,1 +1,0 @@
-rtps_relay_unit_tests
