+++ tests/security/security_tests.lst
@@ -47,7 +47,8 @@ tests/security/attributes/run_test.pl unsecure-part-user-data: !NO_BUILT_IN_TOPI
 tests/security/TopicWithoutPermissions/run_test.pl
 tests/security/SingleParticipantWithSecurity/run_test.pl: !DDS_NO_OWNERSHIP_PROFILE
 
-tests/DCPS/RtpsRelay/Unit/run_test.pl: !DCPS_MIN CXX11 !NO_BUILT_IN_TOPICS !OPENDDS_SAFETY_PROFILE
+tests/run_test.pl unit-tests/tools/rtpsrelay/lib/UnitTests_Name: !DCPS_MIN CXX11 !NO_BUILT_IN_TOPICS !OPENDDS_SAFETY_PROFILE
+tests/run_test.pl unit-tests/tools/rtpsrelay/lib/UnitTests_PartitionIndex: !DCPS_MIN CXX11 !NO_BUILT_IN_TOPICS !OPENDDS_SAFETY_PROFILE
 tests/DCPS/RtpsRelay/Smoke/run_test.pl: !DCPS_MIN CXX11 RTPS !NO_BUILT_IN_TOPICS !OPENDDS_SAFETY_PROFILE RAPIDJSON
 tests/DCPS/RtpsRelay/Smoke/run_test.pl secure: !DCPS_MIN CXX11 RTPS !NO_BUILT_IN_TOPICS !OPENDDS_SAFETY_PROFILE RAPIDJSON
 tests/DCPS/RtpsRelay/Smoke/run_test.pl join: !DCPS_MIN CXX11 RTPS !NO_BUILT_IN_TOPICS !OPENDDS_SAFETY_PROFILE RAPIDJSON
